+++ packages/components/src/components/multi-select/multi-select/multi-select.tsx
@@ -179,6 +179,13 @@ export class MultiSelect {
     this.updateSelectedOption(e.target);
   }
 
+  @Listen('internalOptgroupUpdate')
+  public optgroupUpdateHandler(e: Event): void {
+    e.stopPropagation();
+    this.updateOptions();
+    syncMultiSelectChildrenProps([...this.multiSelectOptions, ...this.multiSelectOptgroups], this.theme);
+  }
+
   @Watch('value')
   public onValueChange(): void {
     this.setFormValue(this.value);