+++ packages/components/CHANGELOG.md
@@ -19,6 +19,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 - `AI-Tag`: New component to mark AI-generated/AI-modified content
   ([#4236](https://github.com/porsche-design-system/porsche-design-system/pull/4236))
 
+### Fixed
+
+- `Multi Select`, `Select`: not correctly updating when options within `p-optgroup` change dynamically
+  ([#4278](https://github.com/porsche-design-system/porsche-design-system/pull/4279))
+
 ## [3.34.0] - 2026-04-09
 
 ## [3.34.0-rc.0] - 2026-04-01